refactor: Move authorization API call to LoginToSpotify button
