+++ src/auth/authCodeWithPkce.ts
@@ -1,3 +1,23 @@
+function generateCodeVerifier(length: number) {
+	let text = ''
+	const possible =
+		'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789'
+
+	for (let i = 0; i < length; i++) {
+		text += possible.charAt(Math.floor(Math.random() * possible.length))
+	}
+	return text
+}
+
+async function generateCodeChallenge(codeVerifier: string) {
+	const data = new TextEncoder().encode(codeVerifier)
+	const digest = await window.crypto.subtle.digest('SHA-256', data)
+	return btoa(String.fromCharCode.apply(null, [...new Uint8Array(digest)]))
+		.replace(/\+/g, '-')
+		.replace(/\//g, '_')
+		.replace(/=+$/, '')
+}
+
 export async function redirectToAuthCodeFlow(clientId: string) {
 	const verifier = generateCodeVerifier(128)
 	const challenge = await generateCodeChallenge(verifier)
@@ -33,24 +53,4 @@ export async function getAccessToken(clientId: string, code: string) {
 
 	const { access_token } = await result.json()
 	return access_token
-}
-
-function generateCodeVerifier(length: number) {
-	let text = ''
-	const possible =
-		'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789'
-
-	for (let i = 0; i < length; i++) {
-		text += possible.charAt(Math.floor(Math.random() * possible.length))
-	}
-	return text
-}
-
-async function generateCodeChallenge(codeVerifier: string) {
-	const data = new TextEncoder().encode(codeVerifier)
-	const digest = await window.crypto.subtle.digest('SHA-256', data)
-	return btoa(String.fromCharCode.apply(null, [...new Uint8Array(digest)]))
-		.replace(/\+/g, '-')
-		.replace(/\//g, '_')
-		.replace(/=+$/, '')
 }
